+++ ktor-http/ktor-http-cio/common/test/io/ktor/tests/http/cio/HeaderParserTest.kt
@@ -79,6 +79,54 @@ class HeaderParserTest {
     }
 
     @Test
+    fun parseHeadersWithMultipleValuesSeparatedWithComma(): Unit = test {
+        val encodedHeaders = """
+            name:value1,value2,value3
+        """.trimIndent() + "\r\n\r\n"
+        val channel = ByteReadChannel(encodedHeaders)
+        val headers = parseHeaders(channel)
+
+        try {
+            assertEquals(1, headers.size)
+            assertEquals("value1,value2,value3", headers["name"].toString())
+            assertEquals(listOf("value1,value2,value3"), headers.getAll("name").map { it.toString() }.toList())
+        } finally {
+            headers.release()
+        }
+    }
+
+    @Test
+    fun parseHeadersWithEmptyValue(): Unit = test {
+        val encodedHeaders = """
+            name:
+        """.trimIndent() + "\r\n\r\n"
+        val channel = ByteReadChannel(encodedHeaders)
+        val headers = parseHeaders(channel)
+        try {
+            assertEquals(1, headers.size)
+            assertEquals("", headers["name"].toString())
+            assertEquals(headers.getAll("name").map { it.toString() }.toList(), listOf(""))
+        } finally {
+            headers.release()
+        }
+    }
+
+    @Test
+    fun parseHeadersWithMultipleEmptyValues(): Unit = test {
+        val encodedHeaders = """
+            name: ,,,
+        """.trimIndent() + "\r\n\r\n"
+        val channel = ByteReadChannel(encodedHeaders)
+        val headers = parseHeaders(channel)
+        try {
+            assertEquals(1, headers.size)
+            assertEquals(",,,", headers["name"].toString())
+        } finally {
+            headers.release()
+        }
+    }
+
+    @Test
     fun parseHeadersSpaceAfterHeaderNameShouldBeProhibited(): Unit = test {
         val encodedHeaders = """
             name :value