+++ ktor-network/posix/src/io/ktor/network/selector/WorkerSelectorManager.kt
@@ -8,12 +8,8 @@ import kotlinx.coroutines.*
 import kotlinx.io.*
 import kotlin.coroutines.*
 
-@OptIn(ExperimentalCoroutinesApi::class)
-internal class WorkerSelectorManager : SelectorManager {
-    @OptIn(DelicateCoroutinesApi::class)
-    private val selectorContext = newSingleThreadContext("WorkerSelectorManager")
-    private val job = Job()
-    override val coroutineContext: CoroutineContext = selectorContext + job
+internal class WorkerSelectorManager(context: CoroutineContext) : SelectorManager {
+    override val coroutineContext: CoroutineContext = context + CoroutineName("selector")
 
     private val selector = SelectorHelper()
 
@@ -39,7 +35,5 @@ internal class WorkerSelectorManager : SelectorManager {
 
     override fun close() {
         selector.requestTermination()
-        selectorContext.close()
-        job.cancel()
     }
 }