+++ ktor-server/ktor-server-test-suites/jvm/src/io/ktor/server/testing/suites/HttpServerJvmTestSuite.kt
@@ -399,14 +399,13 @@ abstract class HttpServerJvmTestSuite<TEngine : ApplicationEngine, TConfiguratio
             val response = parseResponse(ch)!!
 
             assertEquals(HttpStatusCode.SwitchingProtocols.value, response.status)
-            assertEquals("Upgrade", response.headers[HttpHeaders.Connection]?.toString())
-            assertEquals("up", response.headers[HttpHeaders.Upgrade]?.toString())
+            assertEquals("Upgrade", response.headers[HttpHeaders.Connection].toString())
+            assertEquals("up", response.headers[HttpHeaders.Upgrade].toString())
 
-            (0 until response.headers.size)
-                .map { response.headers.nameAt(it).toString() }
-                .groupBy { it }.forEach { (name, values) ->
-                    assertEquals(1, values.size, "Duplicate header $name")
-                }
+            response.headers.offsets()
+                .map { response.headers.nameAtOffset(it).toString() }
+                .groupBy { it }
+                .forEach { (name, values) -> assertEquals(1, values.size, "Duplicate header $name") }
 
             outputStream.apply {
                 writePacket {